Feat: Add device-side barrier_all and block-scope shmem APIs (#178)
- Add ShmemBarrierAllThread/Block device functions with coordinator-based cross-PE barrier
- Add ShmemBarrierOnStream host/Python API for non-blocking stream-based barrier
- Extend PutMemNbi/PutNbiSignal APIs to Block scope with vectorized BlockCopy primitive
- Add warp/block scope wrappers to device bitcode for DSL JIT linking
- Optimize P2P signal kernels: replace SeqCst with release-system atomics
- Fix DSL JIT hook error, and build warnings

+++ include/mori/application/transport/rdma/rdma.hpp
@@ -175,7 +175,7 @@ struct RdmaEndpoint {
     } else if (vendorId == RdmaDeviceVendorId::Pensando) {
       return core::ProviderType::PSD;
     } else {
-      printf("unknown vendorId %d", vendorId);
+      printf("unknown vendorId %u", static_cast<uint32_t>(vendorId));
       assert(false);
     }
     return core::ProviderType::Unknown;
